fix: resolve clippy warnings across crates
- Collapse nested if statements in styx-parse and styx-cli
- Merge identical if branches in CLI main()
- Use io::Error::other() instead of io::Error::new(Other, e)
- Add #[allow(dead_code)] to prepared deserializer error rendering code

+++ facet-styx/src/error.rs
@@ -186,13 +186,15 @@ impl fmt::Display for StyxErrorKind {
 }
 
 /// Convert a facet_reflect::Span to a Range<usize>.
+#[allow(dead_code)]
 fn reflect_span_to_range(span: &facet_reflect::Span) -> std::ops::Range<usize> {
-    let start = span.offset as usize;
-    let end = start + span.len as usize;
+    let start = span.offset;
+    let end = start + span.len;
     start..end
 }
 
 /// Trait for rendering errors with ariadne diagnostics.
+#[allow(dead_code)]
 pub trait RenderError {
     /// Render this error with ariadne.
     ///
@@ -224,6 +226,7 @@ impl RenderError for DeserializeError<StyxError> {
     }
 }
 
+#[allow(dead_code)]
 fn build_deserialize_error_report<'a>(
     err: &DeserializeError<StyxError>,
     filename: &'a str,
